requirements.txt
astunparse==1.6.3
beniget==0.4.1
braceexpand==0.1.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
Cython==0.29.30
docker-pycreds==0.4.0
expecttest==0.1.3
future==0.18.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
horovod==0.23.0
idna==3.3
numpy==1.22.4
pathtools==0.1.2
Pillow==9.1.1
ply==3.11
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pybind11==2.9.2
pycparser==2.21
python-dateutil==2.8.2
pythran==0.11.0
PyYAML==6.0
requests==2.28.0
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
timm==0.4.9
torch==1.10.0a0+git14028c3
torchvision==0.11.0a0+fa347eb
types-dataclasses==0.6.5
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.13.1
webdataset==0.1.62
